requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
asyncpg>=0.28.0
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0
python-dotenv>=1.0.0
